feat(api): expose window target option on event APIs
