+++ .changes/event-api-window-label.md
@@ -1,0 +1,5 @@
+---
+"@tauri-apps/api": patch
+---
+
+Expose the window target option on event APIs.
